+++ pages/info/faqs.js
@@ -10,7 +10,7 @@ import { faqs } from './service';
 export default function FAQs() {
 
   return (
-    <section class="container is-large has-text-centered my-6">
+    <section class="container has-text-centered my-6 section notification">
       <SectionTitle name="Frequently Asked Questions" />
       {faqs.map((item) => (
         <Collapsible title={item.question} content={item.answer} />