+++ pages/info/terms.js
@@ -10,7 +10,7 @@ export default function terms() {
 
 
        return (
-         <section class="section has-text-centered">
+         <section class="section notification has-text-centered">
            <SectionTitle name="Terms and Conditions" />
            <div class="container box is-size-5-mobile">
              <div className={styles.terms}>